+++ src/scvi/model/_utils.py
@@ -110,19 +110,7 @@ def parse_device_args(
             UserWarning,
             stacklevel=settings.warnings_stacklevel,
         )
-    elif _accelerator == "mps" and accelerator == "auto":
-        # auto accelerator should not default to mps
-        connector = _AcceleratorConnector(accelerator="cpu", devices=devices)
-        _accelerator = connector._accelerator_flag
-        _devices = connector._devices_flag
-        warnings.warn(
-            "`accelerator` has been automatically set to `cpu` although 'mps' exists. If you wish "
-            "to run on mps backend, use explicitly accelerator='mps' in train function."
-            "In future releases it will become default for mps supported machines.",
-            UserWarning,
-            stacklevel=settings.warnings_stacklevel,
-        )
-    elif _accelerator == "mps" and accelerator != "auto":
+    elif _accelerator == "mps":
         warnings.warn(
             "`accelerator` has been set to `mps`. Not all PyTorch operations are implemented "
             "for this backend, so some models may be slower or may fail on unsupported ops; "